requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
sqlparse==0.5.3
typing_extensions==4.14.1
pytest==8.4.1
pytest_django==4.11.1
